recommender/pyproject.toml
[project]
name = "movier"
version = "0.1.0"
description = "Movier"
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
    "grpcio>=1.67.0",
    "grpcio-reflection==1.67.0",
    "grpcio-tools==1.67.0",
    "polars==1.12.0",
    "psycopg2-binary==2.9.10",
    "scikit-learn==1.5.2",
    "grpcio-health-checking==1.67.1",
]
